requirements.txt
streamlit==1.42.2
pandas==2.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
openai==1.66.3
python-dotenv==1.0.1
